# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Caregiver Connect API",
    "version": "1.0.0",
    "description": "API documentation for Caregiver Connect platform."
  },
  "servers": [
    { "url": "https://your-app-name.onrender.com" }
  ],
  "paths": {
...
    "/auth/status": {
      "get": {
        "summary": "Check authentication status",
        "responses": { "200": { "description": "Authenticated user info" }, "401": { "description": "Not authenticated" } }
      }
    },
...
    "/profiles": {
      "get": {
        "summary": "Get all caregiver profiles",
        "responses": { "200": { "description": "List of profiles" } }
      },
      "post": {
        "summary": "Create caregiver profile",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CaregiverProfile" } } }
        },
        "responses": { "201": { "description": "Profile created" }, "400": { "description": "Validation error" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/bookings": {
      "get": {
        "summary": "Get all bookings for current user",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "List of bookings" }, "401": { "description": "Unauthorized" } }
      },
      "post": {
        "summary": "Create booking",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Booking" } } }
        },
        "responses": { "201": { "description": "Booking created" }, "400": { "description": "Validation error" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/reviews": {
      "post": {
        "summary": "Create review",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Review" } } }
        },
        "responses": { "201": { "description": "Review created" }, "400": { "description": "Validation error" }, "401": { "description": "Unauthorized" } }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CaregiverProfile": {
        "type": "object",
        "properties": {
          "bio": { "type": "string" },
          "yearsOfExperience": { "type": "integer" },
          "hourlyRate": { "type": "number" }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "caregiverId": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" },
          "status": { "type": "string" },
          "notes": { "type": "string" }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "bookingId": { "type": "string" },
          "caregiverId": { "type": "string" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}